Refactor: change the way the data is saved
Previously the data was save in 3 different files: off peak, peak and
night times. All this has been change to a single file by using a
dictionary and the function save_dic_to_file. This is a much efficient
way to do it. Also the file can be loaded as a spreadsheet.